REF: linter.jl part of Dolo module.
➜  Dolo git:(linter) ✗ time julia --depwarn=no -e 'using Dolo;lint("examples/models/rbc_dtcc_iid.yaml")'
error at line 1, column 6 : Invalid model name 'Real Business Cycle' >> Model name should be non-empty string
julia --depwarn=no -e 'using Dolo;lint("examples/models/rbc_dtcc_iid.yaml")'  8.98s user 0.38s system 103% cpu 9.058 total
➜  Dolo git:(linter) ✗ time julia linter/dolo-lint.jl examples/models/rbc_dtcc_iid.yaml
error at line 1, column 6 : Invalid model name 'Real Business Cycle' >> Model name should be non-empty string
julia linter/dolo-lint.jl examples/models/rbc_dtcc_iid.yaml  6.47s user 0.34s system 104% cpu 6.507 total
